// src/generated/api/app/types.ts
export type AppSignal = ExternIO;
// ...
export type Signal =
  | {
      type: "app";
      value: {
        /**
         * The Cell from which the signal was emitted
         */
        cell_id: CellId;
        /**
         * The Zome from which the signal was emitted
         */
        zome_name: ZomeName;
        /**
         * The actual signal that was emitted
         */
        signal: AppSignal;
      };
    }
  | {
      type: "app_direct";
      value: {
        /**
         * The receiving `CellId`.
         */
        cell_id: CellId;
        /**
         * The agent who sent the signal.
         */
        from_agent: AgentPubKey;
        /**
         * The payload sent by the remote agent.
         */
        signal: Array<number>;
      };
    }
  | { type: "system"; value: SystemSignal };
// ...
export type SystemSignal =
  | { type: "successful_countersigning"; value: EntryHash }
  | { type: "abandoned_countersigning"; value: EntryHash }
  | {
      type: "restore_complete";
      value: {
        /**
         * The cell that finished restoring.
         */
        cell_id: CellId;
      };
    }
  | {
      type: "app_restore_complete";
      value: {
        /**
         * The app that finished restoring.
         */
        installed_app_id: string;
      };
    }
  | {
      type: "restore_failed";
      value: {
        /**
         * The cell that failed to restore.
         */
        cell_id: CellId;
        /**
         * Why the restore is unrecoverable.
         */
        reason: UnrecoverableCellReason;
      };
    };